feat: Add class-validator npm package and update CreateUserDto
This commit adds the class-validator npm package to the project by including it in the package.json and package-lock.json files. It also updates the CreateUserDto in the src/modules/users/dto/create-user.dto.ts file to include validation decorators from the class-validator package. This allows for validation of user input when creating a new user.